backend/requirements.txt
Django==5.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
dj-database-url==2.2.0
whitenoise==6.9.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
requests==2.32.5
pillow==11.2.1
reportlab==4.4.2
